split detect loop and detect loop functions
