+++ 04-javascript-part-1/src/detectLoop.js
@@ -1,0 +1,14 @@
+const detectLoop = (node) => {
+  const nodes = [];
+  let cur = node;
+  while (cur.next()) {
+    if (nodes.indexOf(cur) !== -1) {
+      return true;
+    }
+    nodes.push(cur);
+    cur = cur.next();
+  }
+  return false;
+};
+
+module.exports = detectLoop;
